Refactor: Extract the Linux-specific code to a separate file
