+++ CMakeLists.txt
@@ -32,7 +32,11 @@ else()
 	find_package(PkgConfig REQUIRED)
 	pkg_check_modules(GTK3 REQUIRED gtk+-3.0)
 	pkg_check_modules(WEBKIT2GTK4 REQUIRED webkit2gtk-4.0)
-	target_include_directories(webview PUBLIC ${GTK3_INCLUDE_DIRS} ${WEBKIT2GTK4_INCLUDE_DIRS})
+	target_include_directories(webview PUBLIC
+		${GTK3_INCLUDE_DIRS}
+		${WEBKIT2GTK4_INCLUDE_DIRS}
+		${CMAKE_CURRENT_SOURCE_DIR}/src/linux
+	)
 	target_link_libraries(webview PUBLIC ${GTK3_LIBRARIES} ${WEBKIT2GTK4_LIBRARIES})
 endif()
